+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Theming:
 
 - Completely switched to Figma variables, no need to use plugins anymore
 - Major update for the Slate theme and minor changes for Reshaped theme
+- Updated raised shadow value
 - Updated color tokens generation, closer to Slate theme now
 - [Breaking] Border rgb color tokens got removed (since some border colors are coming with built-in opacity)
 - generateColors support for manually passing dark mode base values
+- [Breaking] --rs-unit-radius-_ turned into --rs-radius-_
+- [Breaking] unit.radius in theme config is now radius
+- [Breaking] heavy font weight is now called extrabold